Minor layout changes, got stuck
Was starting to add and change the layout to support more functionality (adding up to 6 piece masks to a ListView, allowing removal of 1 or all of said piece masks, and solving a puzzle once all 6 piece masks are provided) but got stuck because I realized that I don't have a way to reset all my ToggleButtons when I reset the PieceMask.

Maybe I need to go back to my idea of using a converter with the ToggleButtons and linking them directly to PieceMask? I thought the event triggers thing was pretty elegant but I guess it won't hold up.