Woo! Converter is working!
Fleshed out the converter properly and now we have each ToggleButton properly bound to the piece mask! So now when the piece mask is reset, so are all the ToggleButtons!

Had to use kind of a hacky work-around in the converter to keep track of the piece mask string so that it could be properly decoupled from the MainViewModel. Basically I just copy it into a local variable in every use of the Convert or ConvertBack function so that I have something to reference when the ConvertBack function is called.

Also: Looks like the .gitignore did nothing. That's fine, I figured as much. So:

TODO:
1) Look into properly setting up .gitignore
2) Have the Solve button properly call the CubeSolver library and find a solution once all 6 piece masks are added.