Connected to CubeSolver library!
Woohoo! It works! Solves cubes and errythang! :D

Now all that's left is to add a Readme explaining how to use it and we're all set! Maybe a more visual way of showing the solution? Heck, maybe even a way to display all possible solutions instead of just the first we hit on?

But in the meantime I'll just leave a usage note here:

The solution is provided in an ordered list. Each piece's position in that list indicates where in the final (unfolded) cube that piece lies. The unfolded cube is laid out in a T shape. Pieces 0-3 are the center line (from top to bottom). Piece 4 is on the top-left. Piece 5 is on the top-right. If a piece needs to be flipped and rotated in the final solution, flip the piece horizontally before rotating it! Otherwise it probably won't fit right.