Minor UI Changes - Color and Reset Button
Made it so that the toggle-buttons now toggle between Dark Gray (unchecked) and Dodger Blue (checked) to make it much easier to see the general piece shape. Also made the center of the piece the same color.

Added a Reset button to make it simpler to change the current piece back to a completely empty one.