Added visual representation of solution
Added 6 grids of togglebuttons, essentially copied over from the main togglebutton grid, all with their own converters (so that the local piece mask inside the converter isn't attempting to store multiple piece masks) and bound to their own piece mask which gets set once a solution is found.

The end result is that now you can see the solution in a visual manner and see exactly how the pieces need to be laid out, instead of having to remember which piece is which and then rotating it and flipping it according to text instructions.